+++ src/uenv/parse.h
@@ -59,6 +59,9 @@ util::expected<std::string, parse_error> parse_path(const std::string& in);
 
 util::expected<uenv_label, parse_error> parse_uenv_label(const std::string& in);
 
+util::expected<uenv_description, parse_error>
+parse_uenv_description(const std::string& in);
+
 // pares a namespacesd uenv label
 util::expected<uenv_nslabel, parse_error>
 parse_uenv_nslabel(const std::string& in);