+++ changelog.md
@@ -1,5 +1,8 @@
 # Changelog
 
+- #129 [feature] `uenv image find` and `uenv image ls` do partial match on uenv names
+- #127 [feature] find views, mount point, etc. information using `uenv image inspect` without opening a uenv.
+
 ## 9.1.2
 
 - #125 extend `uenv image add` to support adding uenv that are already in a repo